Move actions to separate files
Every action now need to be put in "moteino_sensors/actions" dir in
a separate file.
File must have name the same as the action name and there
should be a function with action name in it.

Example:
 Action: turn_on_lights
 File: moteino_sensors/actions/turn_on_lights.py
 Contents of the file:
  def turn_on_lights(data, action_config):
    ...
    return True
