+++ moteino_sensors/actions/log.py
@@ -1,0 +1,9 @@
+import logging
+
+
+LOG = logging.getLogger(__name__)
+
+
+def log(data, action_config):
+  LOG.info("Log action for '%s'", data)
+  return True
